inst_seg/requirements.txt
filelock==3.9.0
fsspec==2023.4.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
pillow==10.2.0
sympy==1.12
typing_extensions==4.8.0